+++ tetris/src/pages/report.css
@@ -7,6 +7,7 @@
   font-weight: 700;
   line-height: normal;
   padding-top: 4rem;
+  padding-bottom: 5%;
 }
 
 .report-detail .time-title {
@@ -45,7 +46,7 @@
   font-style: normal;
   font-weight: 400;
   line-height: normal;
-  margin-top: 5%;
+  margin-top: 7%;
 }
 
 .focus-percent {
